# docs/swagger.yaml
definitions:
...
  episode.Episode:
    properties:
      description:
        type: string
      duration_minutes:
        type: integer
      episode_number:
        type: integer
      id:
        type: integer
      name:
        type: string
      release_date:
        type: string
      season_id:
        type: integer
      url:
        type: string
      user_id:
        type: integer
    required:
    - duration_minutes
    - episode_number
    - name
    - release_date
    - url
    type: object
  episodes.EpisodeRequest:
    properties:
      description:
        type: string
      duration_minutes:
        type: integer
      episode_number:
        type: integer
      name:
        type: string
      release_date:
        type: string
      url:
        type: string
    required:
    - duration_minutes
    - episode_number
    - name
    - release_date
    - url
    type: object
  episodes.EpisodeRequestBody:
    properties:
      episodes:
        items:
          $ref: '#/definitions/episodes.EpisodeRequest'
        type: array
    type: object
  movie.Movie:
    properties:
      ageCategory:
        description: one to many
        type: string
      ageCategoryId:
        type: integer
      category:
        description: one to many
        type: string
      categoryId:
        type: integer
      createdAt:
        type: string
      description:
        type: string
      director:
        type: string
      durationMinutes:
        type: integer
      genres:
        description: many to many table
        items:
          type: string
        type: array
      id:
        type: integer
      isFavorite:
        type: boolean
      keywords:
        type: string
      name:
        type: string
      posters:
        description: Maximum of 5 posters for one movie
        items:
          type: string
        type: array
      producer:
        type: string
      seasonCount:
        type: integer
      seriesCount:
        type: integer
      updatedAt:
        type: string
      userId:
        type: string
      watchCount:
        type: integer
      year:
        type: integer
    required:
    - ageCategoryId
    - categoryId
    - description
    - director
    - durationMinutes
    - genres
    - keywords
    - name
    - producer
    - year
    type: object
...
  routes.DefaultMessageResponse:
    properties:
      message:
        type: string
    type: object
  routes.EpisodeRequest:
    properties:
      description:
        type: string
      duration_minutes:
        type: integer
      episode_number:
        type: integer
      name:
        type: string
      release_date:
        type: string
      url:
        type: string
    required:
    - duration_minutes
    - episode_number
    - name
    - release_date
    - url
    type: object
  routes.ManyMoviesResponse:
    properties:
      movies:
        items:
          $ref: '#/definitions/movie.Movie'
        type: array
    type: object
  routes.MovieCreateRequest:
    properties:
      ageCategoryId:
        type: integer
      categoryId:
        type: integer
      description:
        type: string
      director:
        type: string
      durationMinutes:
        type: integer
      genres:
        description: many to many table
        items:
          type: string
        type: array
      keywords:
        type: string
      name:
        type: string
      producer:
        type: string
      year:
        type: integer
    required:
    - ageCategoryId
    - categoryId
    - description
    - director
    - durationMinutes
    - genres
    - keywords
    - name
    - producer
    - year
    type: object
...
  routes.MovieResponse:
    properties:
      movie:
        $ref: '#/definitions/movie.Movie'
    type: object
...
info:
  contact: {}
paths:
...
  /api/episodes/{id}:
    delete:
      description: Deletes an episode with the specified ID
      parameters:
      - description: Episode ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Episode deleted
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "404":
          description: Episode Not Found
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete an episode
    get:
      description: Retrieves an episode with the specified ID
      parameters:
      - description: Episode ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Episode
          schema:
            $ref: '#/definitions/episode.Episode'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "404":
          description: Episode not found
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
      summary: Retrieve an episode
    post:
      description: Creates a single episode for the specified season
      parameters:
      - description: Season ID
        in: path
        name: id
        required: true
        type: integer
      - description: Episode
        in: body
        name: episode
        required: true
        schema:
          $ref: '#/definitions/routes.EpisodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Episode added
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "404":
          description: Season not found
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
      security:
      - ApiKeyAuth: []
      summary: Create episode
    put:
      description: Updates an existing episode
      parameters:
      - description: Episode ID
        in: path
        name: id
        required: true
        type: integer
      - description: Episode
        in: body
        name: episode
        required: true
        schema:
          $ref: '#/definitions/routes.EpisodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Episode updated
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "404":
          description: Episode not found
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
      security:
      - ApiKeyAuth: []
      summary: Update episode
  /api/episodes/{id}/multiple:
    post:
      description: Creates multiple episodes for the specified season
      parameters:
      - description: Season ID
        in: path
        name: id
        required: true
        type: integer
      - description: Episodes
        in: body
        name: episodes
        required: true
        schema:
          $ref: '#/definitions/episodes.EpisodeRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: Multiple Episodes added
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "404":
          description: Season not found
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
      security:
      - ApiKeyAuth: []
      summary: Create episodes
  /api/movies:
    get:
      description: Retrieves movies for a user
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.ManyMoviesResponse'
        "400":
          description: Invalid limit parameter
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
      summary: Get all movies
    post:
      description: Creates a new movie with provided details
      parameters:
      - description: Movie object
        in: body
        name: movie
        required: true
        schema:
          $ref: '#/definitions/routes.MovieCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Movie Created
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "404":
          description: Category not found
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a new movie
  /api/movies/{id}:
    delete:
      description: Deletes a movie with the specified ID
      parameters:
      - description: Movie ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Movie Deleted
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "404":
          description: Movie not found
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete a movie
    get:
      description: Retrieves details of a specific movie including its seasons and
        similar movies
      parameters:
      - description: Movie ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.MovieResponse'
        "400":
          description: Invalid movie ID
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "404":
          description: Movie not found
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
      summary: Get movie details
...
  /api/movies/genre:
    get:
      description: Retrieves movies by genre for a user
      parameters:
      - description: Genre
        in: query
        name: genre
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
      security:
      - ApiKeyAuth: []
      summary: Get movies by genre
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"